% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydatasets.R
\docType{data}
\name{cpg.pval}
\alias{cpg.pval}
\title{An example of user input cpg.pval}
\format{A named vector contains p-values of each probe tested}
\usage{
cpg.pval
}
\description{
An example of user input cpg.pval
}
\keyword{datasets}
