% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydatasets.R
\docType{data}
\name{CpG2Gene}
\alias{CpG2Gene}
\title{An example of user user-supplied mapping between CpGs and genes}
\format{A data frame contains mapping between CpGs and genes}
\usage{
CpG2Gene
}
\description{
An example of user user-supplied mapping between CpGs and genes
}
\keyword{datasets}
