% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSlisttoy.R
\docType{data}
\name{GS.list}
\alias{GS.list}
\title{An example of user input gene sets}
\format{A list contains user input gene set
names and their corresponding genes}
\usage{
GS.list
}
\description{
An example of user input gene sets
}
\keyword{datasets}
