% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{methylGSA shiny app}
\usage{
runExample(run = TRUE)
}
\arguments{
\item{run}{Run the app or not. Default is TRUE}
}
\value{
The shiny app will be opened in a web browser.
}
\description{
This is an interface for Bioconductor package methylGSA.
}
\note{
In order to run the app, the following R/Bioconductor packages needs
to be installed properly: shinycssloaders, DT, ggplot2, 
IlluminaHumanMethylation450kanno.ilmn12.hg19 (if analyzing 450K array)
IlluminaHumanMethylationEPICanno.ilm10b4.hg19 (if analyzing EPIC array)
}
\examples{
## Please note: in this example, the argument run is set to be FALSE in 
## order to pass R CMD check. However, when using the app, users are 
## expected to launch the app by runExample()
runExample(FALSE)
}
