% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylRRA.R
\name{methylRRA}
\alias{methylRRA}
\title{Enrichment analysis after adjusting multiple p-values of
each gene by Robust Rank Aggregation}
\usage{
methylRRA(cpg.pval, array.type = "450K", CpG2Gene = NULL, method = "ORA",
  GS.list = NULL, GS.idtype = "SYMBOL", GS.type = "GO", minsize = 100,
  maxsize = 500)
}
\arguments{
\item{cpg.pval}{A named vector containing p-values of differential
methylation test. Names should be CpG IDs.}

\item{array.type}{A string. Either "450K" or "EPIC". Default is "450K".
This argument will be ignore if CpG2Gene is provided.}

\item{CpG2Gene}{A matrix or data frame with 1st column CpG ID and 2nd
column gene name. Default is NULL.}

\item{method}{A string. "ORA" or "GSEA". Default is "ORA"}

\item{GS.list}{A list. Default is NULL. If there is no input list,
Gene Ontology is used. Entry names are gene sets names, and elements
correpond to genes that gene sets contain.}

\item{GS.idtype}{A string. "SYMBOL", "ENSEMBL", "ENTREZID" or
"REFSEQ". Default is "SYMBOL".}

\item{GS.type}{A string. "GO", "KEGG", or "Reactome". Default is "GO"}

\item{minsize}{An integer. If the number of genes in a gene set is
less than this integer, this gene set is not tested. Default is 100.}

\item{maxsize}{An integer. If the number of genes in a gene set is
greater than this integer, this gene set is not tested. Default is 500.}
}
\value{
A data frame contains gene set tests results.
}
\description{
This function implements enrichment after adjusting
multiple p-values of each gene by Robust Rank Aggregation.
}
\examples{
data(CpG2Genetoy)
data(cpgtoy)
data(GSlisttoy)
GS.list = GS.list[1:10]
res1 = methylRRA(cpg.pval = cpg.pval, CpG2Gene = CpG2Gene,
method = "ORA", GS.list = GS.list)
head(res1)
}
\references{
Kolde, Raivo, et al. Robust rank aggregation for gene
list integration and meta-analysis. Bioinformatics 28.4 (2012): 573-580.

Phipson, B., Maksimovic, J., and Oshlack, A. (2015).
missMethyl: an R package for analysing methylation data from Illuminas
HumanMethylation450 platform. Bioinformatics, btv560.

Yu, Guangchuang, et al. clusterProfiler: an R package for
comparing biological themes among gene clusters. Omics: a journal of
integrative biology 16.5 (2012): 284-287.

Carlson M (2017). org.Hs.eg.db: Genome wide annotation for
Human. R package version 3.5.0.
}
